+++ src/main/java/ua/lviv/iot/services/VaseController.java
@@ -5,29 +5,47 @@ import org.springframework.web.bind.annotation.*;
 import ua.lviv.iot.museum.models.Vase;
 import ua.lviv.iot.second.VaseRepository;
 
+import java.util.LinkedList;
+import java.util.List;
+
 @RestController
 public class VaseController {
 
     @Autowired
     private VaseRepository repository;
 
-    @GetMapping("/vase/{id}")
-    public Vase getVase(@PathVariable Integer id) {
-        return repository.findById(id).get();
+    @RequestMapping(value = "/vase/", method = RequestMethod.GET)
+    public List<Vase> getAllVases() {
+        List<Vase> vases = new LinkedList<>();
+        repository.findAll().forEach(vases::add);
+        if(vases.isEmpty()) {
+            return null;
+        }
+        return vases;
     }
 
-    @PostMapping("/vase")
+    @RequestMapping(value = "/vase/{id}", method = RequestMethod.GET)
+    public Vase getVase(@PathVariable Integer id) {
+        if(repository.findById(id).isPresent()) {
+            return repository.findById(id).get();
+        }
+        return null;
+    }
+
+    @RequestMapping(value = "/vase", method = RequestMethod.POST)
     public Vase postVase(@RequestBody Vase vase) {
         return repository.save(vase);
     }
 
-    @PutMapping("/vase/{id}")
+    @RequestMapping(value = "/vase/{id}", method = RequestMethod.PUT)
     public Vase putVase(@RequestBody Vase vase, @PathVariable Integer id) {
         return repository.save(vase);
     }
 
-    @DeleteMapping("/vase/{id}")
+    @RequestMapping(value = "/vase/{id}", method = RequestMethod.DELETE)
     public void deleteVase(@PathVariable Integer id) {
-        repository.deleteById(id);
+        if(repository.findById(id).isPresent()) {
+            repository.deleteById(id);
+        }
     }
 }
